+++ visao/veiculo/pagAutorizaExcluir.php
@@ -1,5 +1,6 @@
-		<H1> Você realmente deseja excluir o veiculo <?php echo $v->getNome(); ?> ? </H1>
-		
-		<A href="veiculo.php?fun=excluir&conf=sim&id=<?php echo $v->getIdveiculo(); ?>"> SIM </A>
-		
-		<A href="veiculo.php?fun=listar">NÃO</A>
+	<H1 class="titulo"> Você realmente deseja excluir o veiculo <?php echo $v->getNome(); ?> ? </H1>
+
+	<div class="centralizar-div">
+		<A class="btn-confirmar" href="veiculo.php?fun=excluir&conf=sim&id=<?php echo $v->getIdveiculo(); ?>"> SIM </A>
+		<A class="btn-cancelar" href="veiculo.php?fun=listar">NÃO</A>
+	</div>
